move map to utils
