+++ packages/types/src/index.ts
@@ -3,7 +3,7 @@ export { Client } from './client';
 export { Identity, IdentityListener, DEFAULT_IDENTITY_INSTANCE } from './identity';
 export { Event, Payload } from './event';
 export { Options } from './options';
-export { Response, ResponseBody, mapJSONToResponse, mapHttpMessageToResponse, SKIPPED_RESPONSE } from './response';
+export { Response, ResponseBody, SKIPPED_RESPONSE } from './response';
 export { RetryClass } from './retry';
 export { Status } from './status';
 export { Transport, TransportOptions } from './transport';
