+++ packages/utils/src/index.ts
@@ -1,4 +1,4 @@
 export { generateBase36Id, DEVICE_ID_LENGTH } from './base36';
 export { logger } from './logger';
 export { asyncSleep, getGlobalAmplitudeNamespace, isBrowserEnv, isNodeEnv, prototypeJsFix } from './misc';
-export { collectInvalidEventIndices } from './response';
+export { collectInvalidEventIndices, mapHttpMessageToResponse, mapJSONToResponse } from './response';
